#7 CHANGED:  moved demo view into a separate component
- helps to reproduce bug
